add login abd register, add middlwere rols redy -50%

+++ app/Http/Controllers/Admin/PostsController.php
@@ -48,6 +48,8 @@ class PostsController extends Controller
 
        public function destroy(Post $Post)
        {
+          $this->authorize('post_delete');
+
           $Post->delete();
 
           return redirect('/general_admin/posts');
